Create password_generator.py
Passwords we generally create are easy to hack. We may set the same password for many accounts on different platforms. When those platforms encounter a data breach, our password may be leaked, putting not only our account on that platform in jeopardy but can also be employed to unlock other accounts of ours. To avoid  this we can use this password generator to generate random, difficult to hack, passwords, which are unique.